# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "127.0.0.1:5555"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /attcategory:
    get:
      tags:
      - "attcategory"
      summary: "get all attcategory by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AttcategoryPagination"
    post:
      tags:
      - "attcategory"
      summary: "create attcategory"
      description: "create attcategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create attcategory"
        required: true
        schema:
          $ref: "#/definitions/Attcategory"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "attcategory"
      summary: "update attcategory"
      description: "update attcategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create attcategory"
        required: true
        schema:
          $ref: "#/definitions/Attcategory"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /attcategory/{Id}:
    get:
      tags:
      - "attcategory"
      summary: "get a attcategory by ID"
      description: "get a attcategory by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of attcategory to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Attcategory"
    delete:
      tags:
      - "attcategory"
      summary: "Destroy a attcategory by ID"
      description: "delete a attcategory by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of attcategory to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /attcategorytenant:
    get:
      tags:
      - "attcategorytenant"
      summary: "get all attcategorytenant by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AttcategorytenantPagination"
    post:
      tags:
      - "attcategorytenant"
      summary: "create attcategorytenant"
      description: "create attcategorytenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create attcategorytenant"
        required: true
        schema:
          $ref: "#/definitions/Attcategorytenant"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "attcategorytenant"
      summary: "update attcategorytenant"
      description: "update attcategorytenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create attcategorytenant"
        required: true
        schema:
          $ref: "#/definitions/Attcategorytenant"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /attcategorytenant/{Id}:
    get:
      tags:
      - "attcategorytenant"
      summary: "get a attcategorytenant by ID"
      description: "get a attcategorytenant by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of attcategorytenant to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Attcategorytenant"
    delete:
      tags:
      - "attcategorytenant"
      summary: "Destroy a attcategorytenant by ID"
      description: "delete a attcategorytenant by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of attcategorytenant to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /attendance:
    get:
      tags:
      - "attendance"
      summary: "get all attendance by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AttendancePagination"
    post:
      tags:
      - "attendance"
      summary: "create attendance"
      description: "create attendance"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create attendance"
        required: true
        schema:
          $ref: "#/definitions/Attendance"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "attendance"
      summary: "update attendance"
      description: "update attendance"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create attendance"
        required: true
        schema:
          $ref: "#/definitions/Attendance"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /attendance/{Id}:
    get:
      tags:
      - "attendance"
      summary: "get a attendance by ID"
      description: "get a attendance by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of attendance to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Attendance"
    delete:
      tags:
      - "attendance"
      summary: "Destroy a attendance by ID"
      description: "delete a attendance by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of attendance to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /bankaccount:
    get:
      tags:
      - "bankaccount"
      summary: "get all bankaccount by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/BankaccountPagination"
    post:
      tags:
      - "bankaccount"
      summary: "create bankaccount"
      description: "create bankaccount"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create bankaccount"
        required: true
        schema:
          $ref: "#/definitions/Bankaccount"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "bankaccount"
      summary: "update bankaccount"
      description: "update bankaccount"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create bankaccount"
        required: true
        schema:
          $ref: "#/definitions/Bankaccount"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /bankaccount/{Id}:
    get:
      tags:
      - "bankaccount"
      summary: "get a bankaccount by ID"
      description: "get a bankaccount by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of bankaccount to return"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Bankaccount"
    delete:
      tags:
      - "bankaccount"
      summary: "Destroy a bankaccount by ID"
      description: "delete a bankaccount by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of bankaccount to delete"
        required: true
        type:  "string" 
        format:  "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
definitions:
...
  Attcategory:
    type: "object"
    properties:
    
      id:
        type: "string"
        description: "Id"
        format: "string"
      
      holiday_code:
        type: "string"
        description: "编号"
        format: "string"
      
      holiday_name:
        type: "string"
        description: "名称"
        format: "string"
      
      calc_type:
        type: "integer"
        description: "计算类别"
        format: "int64"
      
      is_spec_calc:
        type: "integer"
        description: "特殊计算规则"
        format: "int64"
      
      is_on_slip:
        type: "integer"
        description: "是否显示在工资单上"
        format: "int64"
      
      att_field_name:
        type: "string"
        description: "考勤存储的字段"
        format: "string"
      
      field_name:
        type: "string"
        description: "存储字段"
        format: "string"
      
      formula:
        type: "string"
        description: "公式"
        format: "string"
      
      formula_remark:
        type: "string"
        description: "公式备注"
        format: "string"
      
      sql_formula:
        type: "string"
        description: "SQL公式，没有用的字段"
        format: "string"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  AttcategoryPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Attcategory"
  Attcategorytenant:
    type: "object"
    properties:
    
      holiday_category_id:
        type: "string"
        description: "Id"
        format: "string"
      
      tenant_id:
        type: "string"
        description: "租户编号"
        format: "string"
      
      holiday_code:
        type: "string"
        description: "编号"
        format: "string"
      
      tenant_holiday_name:
        type: "string"
        description: "名称"
        format: "string"
      
      tenant_formula:
        type: "string"
        description: "公式"
        format: "string"
      
      tenant_formula_remark:
        type: "string"
        description: "公式备注"
        format: "string"
      
      is_on_slip:
        type: "integer"
        description: "是否显示在工资单上"
        format: "int64"
      
      is_used:
        type: "integer"
        description: "是否启用"
        format: "int64"
      
      sqltenant_formula:
        type: "string"
        description: "SQL公式"
        format: "string"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  AttcategorytenantPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Attcategorytenant"
  Attendance:
    type: "object"
    properties:
    
      id:
        type: "string"
        description: "编号"
        format: "string"
      
      tenant_id:
        type: "string"
        description: "租户编号"
        format: "string"
      
      person_id:
        type: "string"
        description: "雇员编号"
        format: "string"
      
      month_id:
        type: "string"
        description: "月份编号"
        format: "string"
      
      chi_dao_count:
        type: "number"
        description: "迟到次数"
        format: "float"
      
      chi_dao_num:
        type: "number"
        description: "迟到时间"
        format: "float"
      
      zaotui_count:
        type: "number"
        description: "早退次数"
        format: "float"
      
      zaotui_num:
        type: "number"
        description: "早退时间"
        format: "float"
      
      tiaoxiu_num:
        type: "number"
        description: "调休"
        format: "float"
      
      shi_jia_num:
        type: "number"
        description: "事假"
        format: "float"
      
      sick_num:
        type: "number"
        description: "病假"
        format: "float"
      
      bu_rv_jia_num:
        type: "number"
        description: "哺乳假"
        format: "float"
      
      pei_chan_jia_num:
        type: "number"
        description: "陪产假"
        format: "float"
      
      chan_jia_num:
        type: "number"
        description: "产假"
        format: "float"
      
      hun_jia_num:
        type: "number"
        description: "婚假"
        format: "float"
      
      sang_jia_num:
        type: "number"
        description: "丧假"
        format: "float"
      
      kuanggong_num:
        type: "number"
        description: "旷工"
        format: "float"
      
      kong_bai_num:
        type: "number"
        description: "考勤空白"
        format: "float"
      
      dai_xin_shi_jia_num:
        type: "number"
        description: "带薪事假"
        format: "float"
      
      plan_work_day:
        type: "number"
        description: "应上班天数"
        format: "float"
      
      actual_work_day:
        type: "number"
        description: "实际出勤天数"
        format: "float"
      
      annual_leave_num:
        type: "number"
        description: "年假"
        format: "float"
      
      f_enabled_mark:
        type: "number"
        description: "有效标志0否1是"
        format: "float"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  AttendancePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Attendance"
  Bankaccount:
    type: "object"
    properties:
    
      id:
        type: "string"
        description: "编号"
        format: "string"
      
      person_id:
        type: "string"
        description: "员工编号"
        format: "string"
      
      item_index:
        type: "integer"
        description: "序号"
        format: "int64"
      
      bank_name:
        type: "string"
        description: "开户行"
        format: "string"
      
      bank_branch:
        type: "string"
        description: "开户分行"
        format: "string"
      
      account_no:
        type: "string"
        description: "账号"
        format: "string"
      
      account_name:
        type: "string"
        description: "账号姓名"
        format: "string"
      
      f_enabled_mark:
        type: "integer"
        description: "有效标志0否1是"
        format: "int64"
      
      f_create_user_id:
        type: "string"
        description: "创建人ID"
        format: "string"
      
      f_modify_user_name:
        type: "string"
        description: "编辑人"
        format: "string"
      
      f_modify_user_id:
        type: "string"
        description: "编辑人ID"
        format: "string"
      
      f_delete_mark:
        type: "integer"
        description: "删除标记0否1是"
        format: "int64"
      
      f_sort_code:
        type: "integer"
        description: "排序码"
        format: "int64"
      
      f_modify_date:
        type: "string"
        description: "编辑日期"
        format: "date-time"
      
      f_create_user_name:
        type: "string"
        description: "创建人"
        format: "string"
      
      f_create_date:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      f_description:
        type: "string"
        description: "备注"
        format: "string"
  BankaccountPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Bankaccount"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...